Day 5 part 2
I confess that the solution is a bit
"fuzzy" - it will do attempts until a fixed
state for each section is reached - but I
did not verify if there could be complicated
arrangements that would be stuck in a loop.

The retries, and "proceed to next" way
I set things up did work (for my
input set, at least).